Got this code to run correctly
Issues were with not deallocating local varaiables and consequently
grabbing bad data for the return address.

Current issue: The way I've fixed that is by deallocating local vars
when we return, but this won't work for loops because there are multiple allocs
of the same var. Really, I should be doing this when a block ends because that
ends the scope of the variable.